perf: server render cookie consent banner
