+++ apps/web/src/app/_common/cookies/cookie-dialog.tsx
@@ -1,40 +1,42 @@
 'use client'
 
-import { ChevronDownIcon } from '@heroicons/react/24/solid'
-import { VisuallyHidden } from '@radix-ui/react-visually-hidden'
-import { useIsMounted } from '@sushiswap/hooks'
 import {
   Button,
-  Collapsible,
   Dialog,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
   DialogTrigger,
   LinkExternal,
   Separator,
-  Switch,
   classNames,
 } from '@sushiswap/ui'
+import dynamic from 'next/dynamic'
 import { type ReactElement, useCallback, useEffect, useState } from 'react'
 import { announceCookieChange } from './announce-cookie-change'
+import {
+  type CookieType,
+  type ManageAction,
+  alwaysEnabledCookieTypes,
+  cookieTypes,
+} from './cookie-types'
+
+const ManageCookieDialog = dynamic(() =>
+  import('./manage-cookie-dialog').then((module) => module.ManageCookieDialog),
+)
 
 type BaseAction = 'accept' | 'reject' | 'manage'
 
-function BaseCookieDialog({
+function BaseCookieBanner({
   onAction,
 }: { onAction: (action: BaseAction) => void }) {
   const [isExpanded, setIsExpanded] = useState(false)
+
   return (
-    <DialogContent
-      hideClose
-      className="md:min-w-[720px] !left-[unset] !top-[unset] !bottom-0 md:!right-0 !translate-x-[0%] md:!translate-x-[-50px] md:!translate-y-[-50px]"
-      onOpenAutoFocus={(e) => e.preventDefault()}
+    <section
+      aria-labelledby="cookie-policy-title"
+      className="fixed bottom-0 right-0 z-50 grid w-full gap-4 rounded-t-2xl bg-gray-100 p-6 shadow-lg dark:bg-slate-800 black:bg-secondary md:right-[50px] md:bottom-[50px] md:w-[720px] md:rounded-2xl"
     >
-      <VisuallyHidden>
-        <DialogTitle>Cookie Policy</DialogTitle>
-      </VisuallyHidden>
+      <h2 id="cookie-policy-title" className="sr-only">
+        Cookie Policy
+      </h2>
       <p
         className={classNames(
           'text-sm',
@@ -79,149 +81,11 @@ function BaseCookieDialog({
           Reject all non-essential cookies
         </Button>
       </div>
-    </DialogContent>
+    </section>
   )
 }
 
-const cookieTypes = [
-  'essential',
-  'functional',
-  'analytical',
-  'google',
-  'hotjar',
-] as const
-
-export type CookieType = (typeof cookieTypes)[number]
-
-type ManageAction =
-  | {
-      type: 'confirm'
-    }
-  | {
-      type: 'set'
-      cookieType: CookieType
-      enabled: boolean
-    }
-  | {
-      type: 'reject'
-    }
-
-function ManageCookieDialog({
-  cookieSet,
-  onAction,
-}: { cookieSet: Set<CookieType>; onAction: (action: ManageAction) => void }) {
-  const [isAnalyticsOpen, setIsAnalyticsOpen] = useState(false)
-  return (
-    <>
-      <DialogContent
-        hideClose
-        className="!left-[unset] !top-[unset] !bottom-0 md:!right-0 !translate-x-[0%] md:!translate-x-[-50px] md:!translate-y-[-50px] space-y-4"
-      >
-        <DialogHeader>
-          <DialogTitle className="!mr-0">Manage cookie preferences</DialogTitle>
-        </DialogHeader>
-        <div className="[&>*]:flex [&>*]:justify-between [&>*]:items-center space-y-3">
-          <div>
-            <span>Strictly Neccessary Cookies</span>
-            <Switch checked disabled />
-          </div>
-          <Separator />
-          <div>
-            <span>Functional Cookies</span>
-            <Switch
-              checked={cookieSet.has('functional')}
-              onCheckedChange={(enabled) =>
-                onAction({ type: 'set', cookieType: 'functional', enabled })
-              }
-            />
-          </div>
-          <Separator />
-          <div>
-            <span>Analytical Cookies</span>
-            <Switch
-              checked={cookieSet.has('analytical')}
-              onCheckedChange={(enabled) => {
-                onAction({ type: 'set', cookieType: 'analytical', enabled })
-                onAction({ type: 'set', cookieType: 'google', enabled })
-                onAction({ type: 'set', cookieType: 'hotjar', enabled })
-              }}
-            />
-          </div>
-          <div>
-            <div className="flex flex-col gap-2">
-              <button
-                className="flex gap-1 items-center"
-                type="button"
-                onClick={() => setIsAnalyticsOpen(!isAnalyticsOpen)}
-              >
-                <ChevronDownIcon className="w-3 h-3" />
-                <span>Cookies(2)</span>
-              </button>
-              <Collapsible
-                open={isAnalyticsOpen}
-                className="flex flex-col gap-2 pl-4"
-              >
-                <div className="flex gap-1.5 items-center">
-                  <input
-                    type="checkbox"
-                    checked={cookieSet.has('google')}
-                    onChange={(e) => {
-                      onAction({
-                        type: 'set',
-                        cookieType: 'google',
-                        enabled: e.currentTarget.checked,
-                      })
-                      if (!cookieSet.has('analytical')) {
-                        onAction({
-                          type: 'set',
-                          cookieType: 'analytical',
-                          enabled: true,
-                        })
-                      }
-                    }}
-                  />
-                  Google
-                </div>
-                <div className="flex gap-1.5 items-center">
-                  <input
-                    type="checkbox"
-                    checked={cookieSet.has('hotjar')}
-                    onChange={(e) => {
-                      onAction({
-                        type: 'set',
-                        cookieType: 'hotjar',
-                        enabled: e.currentTarget.checked,
-                      })
-                      if (!cookieSet.has('analytical')) {
-                        onAction({
-                          type: 'set',
-                          cookieType: 'analytical',
-                          enabled: true,
-                        })
-                      }
-                    }}
-                  />
-                  HotJar
-                </div>
-              </Collapsible>
-            </div>
-          </div>
-        </div>
-        <DialogFooter className="!justify-start flex flex-wrap gap-3">
-          <Button onClick={() => onAction({ type: 'confirm' })}>Confirm</Button>
-          <Button
-            variant="secondary"
-            onClick={() => onAction({ type: 'reject' })}
-          >
-            Reject all non-essential cookies
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </>
-  )
-}
-
-const alwaysEnabledCookieTypes = ['essential'] as const
+export type { CookieType } from './cookie-types'
 
 export function CookieDialog({
   defaultOpen,
@@ -229,8 +93,6 @@ export function CookieDialog({
 }: { defaultOpen: boolean; children: ReactElement }) {
   const [open, setOpen] = useState(defaultOpen)
   const [page, setPage] = useState<'base' | 'manage'>('base')
-
-  const isMounted = useIsMounted()
 
   const [enabledCookieSet, setEnabledCookieSet] = useState<Set<CookieType>>(
     new Set(alwaysEnabledCookieTypes),
@@ -297,19 +159,23 @@ export function CookieDialog({
   }, [onConfirm])
 
   return (
-    <Dialog open={open && isMounted} onOpenChange={setOpen}>
-      <DialogTrigger asChild onClick={() => setPage('manage')}>
-        {children}
-      </DialogTrigger>
+    <>
+      {open && page === 'base' ? (
+        <BaseCookieBanner onAction={onBaseAction} />
+      ) : null}
 
-      {page === 'base' ? (
-        <BaseCookieDialog onAction={onBaseAction} />
-      ) : (
-        <ManageCookieDialog
-          cookieSet={enabledCookieSet}
-          onAction={onManageAction}
-        />
-      )}
-    </Dialog>
+      <Dialog open={open && page === 'manage'} onOpenChange={setOpen}>
+        <DialogTrigger asChild onClick={() => setPage('manage')}>
+          {children}
+        </DialogTrigger>
+
+        {page === 'manage' ? (
+          <ManageCookieDialog
+            cookieSet={enabledCookieSet}
+            onAction={onManageAction}
+          />
+        ) : null}
+      </Dialog>
+    </>
   )
 }
