+++ apps/web/src/app/_common/cookies/cookie-types.ts
@@ -1,0 +1,24 @@
+export const cookieTypes = [
+  'essential',
+  'functional',
+  'analytical',
+  'google',
+  'hotjar',
+] as const
+
+export type CookieType = (typeof cookieTypes)[number]
+
+export const alwaysEnabledCookieTypes = ['essential'] as const
+
+export type ManageAction =
+  | {
+      type: 'confirm'
+    }
+  | {
+      type: 'set'
+      cookieType: CookieType
+      enabled: boolean
+    }
+  | {
+      type: 'reject'
+    }
